WWW Spec changes (#42)
* Add websocket denial responses to a common extensions section

This allows HTTP responses such as 401, authentication required to be
sent to the client by a framework that requires authenticated
websocket connections. It would also allow for redirect responses or
other more exotic framework responses.

* Add a server push message extension

This specifies how a client should indicate that a HTTP/2 server push
should be initiated.