Make WsgiToAsgi check scope type
Currently, we validate that messages are of type "http.request". However
this validation only occurs when the first message is received. This may
be too late for servers (such as hypercorn) which try to determine
support for the lifespan extension.

This change checks that scope["type"] is "http" to provide early
failure.