Clarify aspects of the server push extension (#121)
* Clarify the "acceptance" of a push promise

It isn't accepted directly but rather the client can indicate it
doesn't accept any push promises or it can reject specific ones.

* Clarify how the server should set the authority header

I think this makes sense as the push promise should relate to the
request that has triggered it. I also think as it is a pseudo header
it is the server's responsibility to set it.