Clarify Pseudo header treatment
Specifically that Pseudo headers are not an application concern, but
rather are a concern of the server (as they are specific to the
protocol).

Additionally this requires an ASGI server to create a host
header. Doing so ensures that ASGI applications can continue to assume
a host header exists.

Note there can only be one host header which should be at the start as
per RFC-7230, https://tools.ietf.org/html/rfc7230#section-5.4.