Fix memory leak in Local
The _context_refs array of weak references was only freed
when the whole Local object was deleted: even when the
referenced objects weren't there anymore, the array of
empty refs was still there.

This is a problem in django where Local() objects are kept
at the global scope in multiple places and never cleared.