Making thread_sensitive=True the default
This makes thread_sensitive=True the default behaviour for SyncToAsync.

It's easy to run afoul of thread-sensitive code when dealing with
databases or frameworks like Django, and while the penalty for having
this on by default is just slightly worse performance, the penalty for
having it off is very confusing bugs all over the place - so, to keep
with the Django culture of safety first, it's best to default this on.

Additionally, this allows thread_sensitive to be passed into the
`sync_to_async` decorator, so if people are sure their code is safe and
want to run it in its own thread for performance reasons, they can use
`@sync_to_async(thread_sensitive=False)`.