New sync_to_async tests, one of which triggers a deadlock
In these tests, we are waiting on a mulitprocessing.Event to be set,
with the wait in `sync_to_async()`.

The event is being set in an `asyncio.Task` that sleeps and then calls
set, also in `sync_to_async()`. This allows us to kick off the process
that would `set()` the event in advance of `wait()`ing for the event,
as the event might be triggered by some consumer event.

If both `sync_to_async()` calls are `thread_sensitive=True`, then
before `await sync_to_async(event.set)()` can run `event.set()`,
`sync_to_async(event.wait)()` has consumed the thread.

asgiref.timeout is used to get us out of the deadlock so the tests can
run to completion, and in the case of Timeout we `set()` the event to
release the `wait()`.

Finally, we await on the setting Task regardless of whether or not
there was a timeout.