Decide which current_task method to use at import time based on the Python version.
Follows up on f53fbef.

Saves 1 hasattr call per usage of asgiref.timeout.current_task, and the same on Local.__getattr__/__setattr__ (which depends indirectly on SyncToAsync.get_current_task)

Note that the module level variable used in the timeout module has to be aliased to another name, because `current_task` is already in use by the public function name.